+++ src/main/java/gamePieces/bishops/BlackBishop.java
@@ -6,11 +6,11 @@ import org.academiadecodigo.simplegraphics.pictures.Picture;
 public class BlackBishop extends GamePiecesAbs implements GamePieces {
 
     private final String location = "src/main/resources/BlackBishop.png";
-    private int col;
-    private int row;
+    private double col;
+    private double row;
     private Picture blackBishop;
 
-    public BlackBishop(int col, int row) {
+    public BlackBishop(double col, double row) {
         Picture blackBishop = new Picture(col * cellSize + center, row * cellSize + center, location);
         blackBishop.draw();
         this.col = col;
@@ -22,13 +22,12 @@ public class BlackBishop extends GamePiecesAbs implements GamePieces {
 
     @Override
     public void showPossibleMoves() {
-
     }
 
     @Override
-    public void move(int targetCol, int targetRow) {
-        int moveToCol;
-        int moveToRow;
+    public void move(double targetCol, double targetRow) {
+        double moveToCol;
+        double moveToRow;
 
         if(this.col > targetCol) {
             moveToCol = targetCol *(-1);
@@ -47,8 +46,7 @@ public class BlackBishop extends GamePiecesAbs implements GamePieces {
     }
 
     @Override
-    public void eat(int col, int row) {
-
+    public void eat(double col, double row) {
     }
 
     @Override
@@ -58,12 +56,12 @@ public class BlackBishop extends GamePiecesAbs implements GamePieces {
     }
 
     @Override
-    public int getCol() {
+    public double getCol() {
         return col;
     }
 
     @Override
-    public int getRow() {
+    public double getRow() {
         return row;
     }
 }
