+++ src/main/java/gamePieces/horses/WhiteHorse.java
@@ -7,11 +7,11 @@ import org.academiadecodigo.simplegraphics.pictures.Picture;
 public class WhiteHorse extends GamePiecesAbs implements GamePieces {
 
     private final String location = "src/main/resources/WhiteHorse.png";
-    private int col;
-    private int row;
+    private double col;
+    private double row;
     private Picture whiteHorse;
 
-    public WhiteHorse(int col, int row) {
+    public WhiteHorse(double col, double row) {
         Picture whiteHorse = new Picture(col * cellSize + center, row * cellSize + center, location);
         whiteHorse.draw();
         this.col = col;
@@ -27,9 +27,9 @@ public class WhiteHorse extends GamePiecesAbs implements GamePieces {
     }
 
     @Override
-    public void move(int targetCol, int targetRow) {
-        int moveToCol;
-        int moveToRow;
+    public void move(double targetCol, double targetRow) {
+        double moveToCol;
+        double moveToRow;
 
         if(this.col > targetCol) {
             moveToCol = targetCol *(-1);
@@ -48,7 +48,7 @@ public class WhiteHorse extends GamePiecesAbs implements GamePieces {
     }
 
     @Override
-    public void eat(int col, int row) {
+    public void eat(double col, double row) {
 
     }
 
@@ -58,12 +58,12 @@ public class WhiteHorse extends GamePiecesAbs implements GamePieces {
     }
 
     @Override
-    public int getCol() {
+    public double getCol() {
         return col;
     }
 
     @Override
-    public int getRow() {
+    public double getRow() {
         return row;
     }
 }
