+++ src/main/java/gamePieces/kings/WhiteKing.java
@@ -7,11 +7,11 @@ import org.academiadecodigo.simplegraphics.pictures.Picture;
 public class WhiteKing extends GamePiecesAbs implements GamePieces {
 
     private final String location = "src/main/resources/WhiteKing.png";
-    private int col;
-    private int row;
+    private double col;
+    private double row;
     private Picture whiteKing;
 
-    public WhiteKing (int col, int row) {
+    public WhiteKing (double col, double row) {
         Picture whiteKing = new Picture(col* cellSize + center, row * cellSize + center, location);
         whiteKing.draw();
         this.col= col;
@@ -25,9 +25,9 @@ public class WhiteKing extends GamePiecesAbs implements GamePieces {
     }
 
     @Override
-    public void move(int targetCol, int targetRow) {
-        int moveToCol;
-        int moveToRow;
+    public void move(double targetCol, double targetRow) {
+        double moveToCol;
+        double moveToRow;
 
         if(this.col > targetCol) {
             moveToCol = targetCol *(-1);
@@ -46,7 +46,7 @@ public class WhiteKing extends GamePiecesAbs implements GamePieces {
     }
 
     @Override
-    public void eat(int col, int row) {
+    public void eat(double col, double row) {
 
     }
 
@@ -56,12 +56,12 @@ public class WhiteKing extends GamePiecesAbs implements GamePieces {
     }
 
     @Override
-    public int getCol() {
+    public double getCol() {
         return col;
     }
 
     @Override
-    public int getRow() {
+    public double getRow() {
         return row;
     }
 }
