+++ src/main/java/gamePieces/pawns/BlackPawn.java
@@ -7,8 +7,8 @@ import org.academiadecodigo.simplegraphics.pictures.Picture;
 public class BlackPawn extends GamePiecesAbs implements GamePieces {
 
     private final String location = "src/main/resources/BlackPawn.png";
-    private int col;
-    private int row;
+    private double col;
+    private double row;
     private Picture blackPawn;
 
     public BlackPawn(int col, int row) {
@@ -27,9 +27,9 @@ public class BlackPawn extends GamePiecesAbs implements GamePieces {
     }
 
     @Override
-    public void move(int targetCol, int targetRow) {
-        int moveToCol;
-        int moveToRow;
+    public void move(double targetCol, double targetRow) {
+        double moveToCol;
+        double moveToRow;
 
         if(this.col > targetCol) {
             moveToCol = targetCol *(-1);
@@ -48,7 +48,7 @@ public class BlackPawn extends GamePiecesAbs implements GamePieces {
     }
 
     @Override
-    public void eat(int col, int row) {
+    public void eat(double col, double row) {
 
     }
 
@@ -58,12 +58,12 @@ public class BlackPawn extends GamePiecesAbs implements GamePieces {
     }
 
     @Override
-    public int getCol() {
+    public double getCol() {
         return col;
     }
 
     @Override
-    public int getRow() {
+    public double getRow() {
         return row;
     }
 }
