+++ src/main/java/gamePieces/quens/BlackQueen.java
@@ -6,11 +6,11 @@ import org.academiadecodigo.simplegraphics.pictures.Picture;
 public class BlackQueen extends GamePiecesAbs implements GamePieces {
 
     private final String location = "src/main/resources/BlackQueen.png";
-    private int col;
-    private int row;
+    private double col;
+    private double row;
     private Picture blackQueen;
 
-    public BlackQueen(int col, int row) {
+    public BlackQueen(double col, double row) {
         Picture blackQueen = new Picture(col * cellSize + center, row * cellSize + center, location);
         blackQueen.draw();
         this.col = col;
@@ -24,9 +24,9 @@ public class BlackQueen extends GamePiecesAbs implements GamePieces {
     }
 
     @Override
-    public void move(int targetCol, int targetRow) {
-        int moveToCol;
-        int moveToRow;
+    public void move(double targetCol, double targetRow) {
+        double moveToCol;
+        double moveToRow;
 
         if(this.col > targetCol) {
             moveToCol = targetCol *(-1);
@@ -45,7 +45,7 @@ public class BlackQueen extends GamePiecesAbs implements GamePieces {
     }
 
     @Override
-    public void eat(int col, int row) {
+    public void eat(double col, double row) {
 
     }
 
@@ -55,12 +55,12 @@ public class BlackQueen extends GamePiecesAbs implements GamePieces {
     }
 
     @Override
-    public int getCol() {
+    public double getCol() {
         return col;
     }
 
     @Override
-    public int getRow() {
+    public double getRow() {
         return row;
     }
 }
