+++ src/main/java/gamePieces/towers/BlackTower.java
@@ -7,11 +7,11 @@ import org.academiadecodigo.simplegraphics.pictures.Picture;
 public class BlackTower extends GamePiecesAbs implements GamePieces {
 
     private final String location = "src/main/resources/BlackTower.png";
-    private int col;
-    private int row;
+    private double col;
+    private double row;
     private Picture blackTower;
 
-    public BlackTower(int col, int row) {
+    public BlackTower(double col, double row) {
         Picture blackTower = new Picture(col * cellSize + center, row * cellSize + center, location);
         blackTower.draw();
         this.col = col;
@@ -28,9 +28,9 @@ public class BlackTower extends GamePiecesAbs implements GamePieces {
     }
 
     @Override
-    public void move(int targetCol, int targetRow) {
-        int moveToCol;
-        int moveToRow;
+    public void move(double targetCol, double targetRow) {
+        double moveToCol;
+        double moveToRow;
 
         if(this.col > targetCol) {
             moveToCol = targetCol *(-1);
@@ -49,7 +49,7 @@ public class BlackTower extends GamePiecesAbs implements GamePieces {
     }
 
     @Override
-    public void eat(int col, int row) {
+    public void eat(double col, double row) {
 
     }
 
@@ -59,12 +59,12 @@ public class BlackTower extends GamePiecesAbs implements GamePieces {
     }
 
     @Override
-    public int getCol() {
+    public double getCol() {
         return col;
     }
 
     @Override
-    public int getRow() {
+    public double getRow() {
         return row;
     }
 }
